+++ src/main/java/com/home_rental_solution/ms_anfitriones/model/Anfitriones.java
@@ -14,7 +14,7 @@ public class Anfitriones {
 //se quitan anotaciones como @NotNull
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer idAnfitrion;
+    private Long idAnfitrion;
 
     @Column(nullable = false, length = 120)
     private String nombre;
